requirements.txt
# Core Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# LLM (Groq — sub-second Llama 3 inference)
groq==0.12.0

# Google Calendar API
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# HTTP Client (for Sarvam API calls)
requests==2.32.3

# Utilities
python-dotenv==1.0.1
